translation-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
googletrans==3.1.0a0
gunicorn==21.2.0 